feat: refactor task state management and implement getSingle task functionality
