+++ src/tasks/constants/states.ts
@@ -1,0 +1,1 @@
+export const states = ['completed', 'pending', 'overdue'] as const;
